test(federated): move the hint-publication test to a companion file (#313)
PR #362 review P3: adding the test had pushed parquet_source_test.go to
518 lines, over coding-standard.md's 500-line cap. Mechanical move (sed
discipline: body untouched), back to 486; the new file carries only the
imports it needs.
